Block & Regen fixes, Enemy Actions, and Clarity
Clarity:
Added [Pre-Strike] and [Post-Strike] next to name of the one being attacked and a space above it.
Replaced [Enter anything to continue] with [Press Enter to continue]
Replaced [An enemy appears] with [A slime appears] since the slime is only enemy as of now

Enemies can now Block and Do Nothing as well as attack, as the player can

Fixed:
Block: Defense can no longer descend past 0; the amount it falls past 0 is applied to health, then defense is set back to 0.

Regen: Health can no longer surpass what it originally was. If the regen application would make it do so, health will instead be set to the max. If not, regen is applied as per usual.
Fixed the end conditions and added victory & failure messages

Other:
Fixed repetitive if statements for Warrior specialties
Added  + " HP" where it should have been after the enemy's health text while it's being attacked

Rebalanced Berserker and Mage specialties